forge formated

+++ src/EventLogger.sol
@@ -2,7 +2,6 @@
 pragma solidity ^0.8.13;
 
 contract EventLogger {
-
     struct Log {
         string eventType;
         uint256 timestamp;
@@ -38,29 +37,17 @@ contract EventLogger {
         emit SecurityEvent(eventType, timestamp, image, commandLine, pid, user, integrityLevel);
     }
 
-    function getLogCount() public view returns (uint) {
+    function getLogCount() public view returns (uint256) {
         return logs.length;
     }
 
-    function getLog(uint index) public view returns (
-        string memory,
-        uint256,
-        string memory,
-        string memory,
-        string memory,
-        string memory,
-        string memory
-    ) {
+    function getLog(uint256 index)
+        public
+        view
+        returns (string memory, uint256, string memory, string memory, string memory, string memory, string memory)
+    {
         require(index < logs.length, "Index out of bounds");
         Log memory log = logs[index];
-        return (
-            log.eventType,
-            log.timestamp,
-            log.image,
-            log.commandLine,
-            log.pid,
-            log.user,
-            log.integrityLevel
-        );
+        return (log.eventType, log.timestamp, log.image, log.commandLine, log.pid, log.user, log.integrityLevel);
     }
 }
